Add files via upload
#TO DETERMINE IF THE NUMBER IS POSITIVE OR NEGATIVE
From this program I have learn about first to Asks the user to enter a number,than read the number which we have entered than check if the number is positive or negative than tell the user what type of the number it is. 
This program uses basic conditions to determine and display whether the number is positive or negative.