Add files via upload
//Polymorphism

Run-time Polymorphism: This is achieved using virtual functions. At runtime, the correct method, say sound(), is invoked based on the actual class of the object instantiated (Bird or Lion) though you are using a pointer to base class, Animal*.

Compile time Polymorphism: This is achieved by function overloading. The correct function, either display(string) or display(float), is chosen at compile time based on the type of the argument.

Memory Management: new is used to dynamically allocate memory for objects and delete is used for freeing that memory once it is done.