Add files via upload
//Implement Inheritance

Inheritance:

Motorcycle is a derived class which inherits from the base class Vehicle. This means that methods and properties defined in Vehicle can be used in Motorcycle, unless they are overlaid or hidden.
Constructor Chaining:

When a Motorcycle is instantiated, the constructor of Vehicle is automatically executed (as Motorcycle derives from Vehicle) followed by the constructor of Motorcycle.
Method Access:

myMotorcycle can invokes methods from both Vehicle: start() and Motorcycle: rev(), illustrating the ability of an instance of a derived class to invoke methods from both its own class and its base class.